third_party/blink/web_tests/fast/overflow/overflow-clamp-after-visible-rect-resize-expected.txt

Ensure that overflow clamping takes into account the container and content size that are applied in the same animation frame.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS scroller.scrollTop is 200
PASS scroller.scrollTop is 200
PASS successfullyParsed is true

TEST COMPLETE
